gateways/hermes/clawke/requirements.txt
websockets>=12.0
PyYAML>=6.0
